+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version ? (Status: DEVELOPMENT)
 * Fix issue #24 (nafsdm-webinterface password resets on upgrade)
+* Add slavestatus page to nafsdm-webinterface.
 
 Version 1.3.1 (Status: RELEASED)
 * Bugfix: Slaves reported incorrect date to nafsdmctl slavestatus (minute instead of month under the "latest connection date" section)